+++ src/main/java/net/soulsweaponry/enchantments/FastHandsEnchantment.java
@@ -9,7 +9,7 @@ import net.soulsweaponry.items.gun.GunItem;
 public class FastHandsEnchantment extends Enchantment {
 
     public FastHandsEnchantment(Rarity weight, EquipmentSlot... slotTypes) {
-        super(weight, EnchantmentTarget.BOW, slotTypes);
+        super(weight, EnchantmentTarget.BREAKABLE, slotTypes);
     }
     
     @Override